Move functions for transaction recovery
to the transaction_recovery.go file
